Added authentication (login, signup...) and flash messages

+++ src/controllers/image.js
@@ -9,7 +9,7 @@ const ctrl = {}
 ctrl.index = async (req, res) => {
     let viewModel = { image: {}, comments: {} };
     const image = await Image.findOne({ filename: {$regex: req.params.image_id} });
-    
+
     if (image) {
         image.views += 1;
         viewModel.image = image;
@@ -47,9 +47,9 @@ ctrl.create = (req, res) => {
                     title: req.body.title,
                     filename: imageUrl + ext,
                     description: req.body.description,
+                    user: req.user.email
                 });
-                
-                const imageSaved = await newImage.save();
+                await newImage.save();
                 res.redirect(`/images/${imageUrl}`)
             } else {
             
@@ -91,14 +91,28 @@ ctrl.comment = async (req, res) => {
 }
 
 ctrl.remove = async (req, res) => {
-    const image = await Image.findOne({ filename: {$regex: req.params.image_id} });
-    
-    if (image) {
-        await fs.unlink(path.resolve(`./src/public/upload/${image.filename}`));
-        await Comment.deleteMany({ image_id: image._id });
-        await image.remove();
-        res.json(true);
+
+    const authUser = req.user;
+    let imageUser;
+
+    if (authUser){
+        imageUser = await Image.findOne({ user: authUser.email})
     }
+    
+    
+    if (imageUser) {
+        
+        const image = await Image.findOne({ filename: {$regex: req.params.image_id} });
+        
+        if (image) {
+            await fs.unlink(path.resolve(`./src/public/upload/${image.filename}`));
+            await Comment.deleteMany({ image_id: image._id });
+            await image.remove();
+            res.json(true);
+        }
+
+    }
+    
 }
 
 module.exports = ctrl;
